Got it to run on my personal computer.
It's been almost a year since I've actively worked on this complex
project. I have since switched to using Gentoo Linux from Windows 10.
Not to mention I have a lot more experience in writing C now and reading
this C++ code gives me a headache. So there is a lot to come to fix the
technical debt that I have created in this project.