Reorganized CMakeLists tree.
After using make for a while I think it's unnecessary to have a bunch of
CMakeLists in each directory. Makes it harder to track the sources. Now
all one needs to do is have a command like tree and two terminals open
to add a source to the main CMakeLists